[Kinetics] Move BlowersMaselRate to its own header file
